[IMP] product_seasonality refactoring
refactor module to add a base module that handles allowed product lists.
Product_seasonality comes to add dates rules on top of this base module
