+++ product_allowed_list/models/__init__.py
@@ -1,0 +1,3 @@
+from . import product_list
+from . import product_list_line
+from . import res_partner
