+++ product_allowed_list/readme/CONFIGURE.rst
@@ -1,0 +1,3 @@
+Go to "Contacts -> Partner" and configure the product lists configuration for the specific partner.
+
+The configuration for products sale is available when `sale_product_list` is installed.
