+++ product_allowed_list/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* Simone Orsi <simone.orsi@camptocamp.com>
+* Julien Coux <julien.coux@camptocamp.com>
+* Damien Crier <damien.crier@camptocamp.com>
