+++ product_allowed_list/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+Define rules for products' list.
+Those products' lists can be used in other modules to restrict product availability in sales, purchase,...
