+++ product_allowed_list/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_product_allowed_list
