+++ product_allowed_list/views/res_partner.xml
@@ -8,14 +8,14 @@
 
             <group name="sale" position="inside">
                 <field
-                    name="seasonal_config_id"
+                    name="product_list_ids"
                     attrs="{'invisible': [('is_company','=',False), ('parent_id','!=',False)]}"
                 />
                 <div
                     colspan="2"
                     attrs="{'invisible': ['|',('is_company','=',True),('parent_id','=',False)]}"
                 >
-                    <p>Product seasonal availability is managed on <button
+                    <p>Product lists are managed on <button
                             name="open_commercial_entity"
                             type="object"
                             string="the parent company"
