+++ product_seasonality/models/__init__.py
@@ -1,4 +1,2 @@
-from . import seasonal_config
-from . import seasonal_config_line
+from . import product_allowed_list_line
 from . import res_company
-from . import res_partner
