+++ product_seasonality/readme/CONFIGURE.rst
@@ -1,4 +1,4 @@
-Go to "Settings -> Companies" and configure the default seasonal configuration for a company.
-Go to "Contacts -> Partner" and configure the seasonal configuration for the specific partner.
+Go to "Settings -> Companies" and configure the default product list configuration for a company.
+Go to "Contacts -> Partner" and configure the product list configuration for the specific partner.
 
-The configuration for products sale is available when `sale_product_seasonality` is installed.
+The configuration for products sale is available when `sale_product_allowed` and/or `sale_product_seasonality` are installed.
