+++ setup/product_allowed_list/odoo/addons/product_allowed_list
@@ -1,0 +1,1 @@
+../../../../product_allowed_list
